create remove_additem
